pyproject.toml
[project]
name = "ado-asana-sync"
version = "0.0.0-dev"
description = "Tool to sync work items and pull requests from Azure DevOps to Asana"
authors = [{name = "Dan Anstis", email = "dan@bsod.co.nz"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "asana>=5.0.3",
    "azure-devops>=7.1.0b3",
    "tinydb>=4.8.0",
    "pytz>=2023.3,<2026.0",
    "azure-monitor-opentelemetry>=1.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.5.1",
    "pytest-cov>=4.1,<7.0",
    "pytest>=7.4,<9.0",
    "mdformat>=0.7.0",
    "mdformat-gfm>=0.3,<0.5",
]

[tool.ruff]
line-length = 127

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "S",   # bandit security
]
ignore = [
    "S608",  # hardcoded-sql-expression - handled with nosec comments
]

[tool.ruff.lint.per-file-ignores]
# Build/quality scripts legitimately use subprocess
"scripts/quality.py" = ["S603", "S607"]
# Test files use hardcoded test tokens which are safe
"tests/**/*.py" = ["S106"]
# Database tests need to catch generic exceptions for constraint violations
"tests/database/test_database.py" = ["B017"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.coverage.run]
relative_files = true
omit = ["tests/*"]
branch = true

[project.scripts]
# Main application
ado-asana-sync = "ado_asana_sync.sync.__main__:main"

# Quality assurance scripts
lint = "scripts.quality:lint"
format = "scripts.quality:format_code"
format-check = "scripts.quality:format_check"
type-check = "scripts.quality:type_check"
test = "scripts.quality:test"
check = "scripts.quality:check_all"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
